moved variant graphs to variantViz.py
