+++ mango-python/bdgenomics/mango/QC.py
@@ -223,227 +223,3 @@ class AlignmentDistribution(object):
 
         return chromosomes
 
-class VariantEffectAlleleFreq(object):
-    """
-    todo: this needs to be modified to calculate bins in Spark so as to reduce Collect size
-    """
-    def __init__(self, ss, variantRDD):
-
-        so_terms = [
-            'intergenic_region',
-            'downstream_gene_variant',
-            'upstream_gene_variant',
-            'sequence_feature',
-            'structural_interaction_variant',
-            'intron_variant',
-            'non_coding_transcript_exon_variant',
-            '3_prime_UTR_variant',
-            '5_prime_UTR_variant',
-            'synonymous_variant',
-            'conservative_inframe_insertion',
-            'conservative_inframe_deletion',
-            'initiator_codon_variant',
-            'protein_protein_contact',
-            'non_coding_transcript_variant',
-            'stop_retained_variant',
-            '5_prime_UTR_premature_start_codon_gain_variant',
-            'disruptive_inframe_deletion',
-            'disruptive_inframe_insertion',
-            'missense_variant',
-            'stop_lost',
-            'start_lost',
-            'splice_donor_variant',
-            'splice_acceptor_variant',
-            'splice_region_variant',
-            'frameshift_variant',
-            'stop_gained']
-        so_to_rank = dict(zip(so_terms, range(1,len(so_terms)+1)))
-        rank_to_so = {v: k for k, v in so_to_rank.iteritems()}
-
-        def most_sig_effect(effectList):
-            return rank_to_so[max( [ so_to_rank[item] for item in effectList] )]
-
-        def jp3(dd1):
-            effectList = [ effect for transcript in dd1['transcriptEffects'] for effect in transcript['effects'] ]
-            mostSigEffect = most_sig_effect(effectList)
-            return mostSigEffect
-
-        z = variantRDD.toDF().rdd.map(lambda w: w.asDict()).map(lambda p: p['annotation']).map(lambda d: (d['attributes']['AF_NFE'], jp3(d)) )
-        self.b_data = z.filter(lambda f: f[1] == 'missense_variant').collect()
-        self.c_data = z.filter(lambda f: f[1] == 'synonymous_variant').collect()
-        self.sc = ss.sparkContext
-
-    def plot(self):
-        plt.rcdefaults()
-        plt.title("Allele Frequency Distribution by Functional Category")
-        fig, ax = plt.subplots(1,figsize=(14, 8))
-        b_data2= [ float(item[0]) for item in self.b_data if item[0] != "."]
-        c_data2= [ float(item[0]) for item in self.c_data if item[0] != "."]
-        ax.hist([b_data2,c_data2],bins=40, normed=1, label=['missense_variant','synonymous_variant'])
-        ax.legend(loc='upper right')
-        ax.set_yscale('log')
-        ax.yaxis.set_major_formatter(ticker.FuncFormatter(lambda y,pos: ('{{:.{:1d}f}}'.format(int(np.maximum(-np.log10(y),0)))).format(y)))
-        ax.yaxis.set_label_text("Percent Variants")
-        ax.xaxis.set_label_text("Allele Frequency")
-        plt.show()
-
-class VariantEffectAlleleFreq2(object):
-    """
-    todo: this needs to be modified to calculate bins in Spark so as to reduce Collect size
-    """
-    def __init__(self, ss, variantRDD):
-
-        so_terms = [
-            'intergenic_region',
-            'downstream_gene_variant',
-            'upstream_gene_variant',
-            'sequence_feature',
-            'structural_interaction_variant',
-            'intron_variant',
-            'non_coding_transcript_exon_variant',
-            '3_prime_UTR_variant',
-            '5_prime_UTR_variant',
-            'synonymous_variant',
-            'conservative_inframe_insertion',
-            'conservative_inframe_deletion',
-            'initiator_codon_variant',
-            'protein_protein_contact',
-            'non_coding_transcript_variant',
-            'stop_retained_variant',
-            '5_prime_UTR_premature_start_codon_gain_variant',
-            'disruptive_inframe_deletion',
-            'disruptive_inframe_insertion',
-            'missense_variant',
-            'stop_lost',
-            'start_lost',
-            'splice_donor_variant',
-            'splice_acceptor_variant',
-            'splice_region_variant',
-            'frameshift_variant',
-            'stop_gained']
-        so_to_rank = dict(zip(so_terms, range(1,len(so_terms)+1)))
-        rank_to_so = {v: k for k, v in so_to_rank.iteritems()}
-
-        def most_sig_effect(effectList):
-            return rank_to_so[max( [ so_to_rank[item] for item in effectList] )]
-
-        def jp3(dd1):
-            effectList = [ effect for transcript in dd1['transcriptEffects'] for effect in transcript['effects'] ]
-            mostSigEffect = most_sig_effect(effectList)
-            return mostSigEffect
-
-
-        z = variantRDD.toDF().rdd.map(lambda w: w.asDict()).map(lambda p: p['annotation']).filter(
-            lambda x: x['attributes']['AF_NFE'] != ".").map(
-            lambda d: (jp3(d), math.floor(float(d['attributes']['AF_NFE'])*10000)/10000)).filter(
-            lambda f: f[0] == 'missense_variant').countByValue()
-        zlist = z.items()
-        zbins = [ item[0][1] for item in zlist ]
-        zweights = [ item[1] for item in zlist ]
-
-        y = variantRDD.toDF().rdd.map(lambda w: w.asDict()).map(lambda p: p['annotation']).filter(
-            lambda x: x['attributes']['AF_NFE'] != ".").map(
-            lambda d: (jp3(d), math.floor(float(d['attributes']['AF_NFE'])*10000)/10000)).filter(
-            lambda f: f[0] == 'synonymous_variant').countByValue()
-        ylist = y.items()
-        ybins = [ item[0][1] for item in ylist ]
-        yweights = [ item[1] for item in ylist ]
-
-        self.z_data = (zbins, zweights)
-        self.y_data = (ybins, yweights)
-
-
-        """
-        z = variantRDD.toDF().rdd.map(lambda w: w.asDict()).map(lambda p: p['annotation']).map(lambda d: (d['attributes']['AF_NFE'], jp3(d)) )
-        self.b_data = z.filter(lambda f: f[1] == 'missense_variant').collect()
-        self.c_data = z.filter(lambda f: f[1] == 'synonymous_variant').collect()
-        """
-        self.sc = ss.sparkContext
-
-    def plot(self):
-        plt.rcdefaults()
-        plt.title("Allele Frequency Distribution by Functional Category")
-        plt.hist([self.z_data[0],self.y_data[0]], weights=[self.z_data[1],self.y_data[1]],bins=20,normed=1)
-        plt.yscale('log')
-        plt.show()
-        """
-        fig, ax = plt.subplots(1,figsize=(14, 8))
-        b_data2= [ float(item[0]) for item in self.b_data if item[0] != "."]
-        c_data2= [ float(item[0]) for item in self.c_data if item[0] != "."]
-        ax.hist([b_data2,c_data2],bins=40, normed=1, label=['missense_variant','synonymous_variant'])
-        ax.legend(loc='upper right')
-        ax.set_yscale('log')
-        ax.yaxis.set_major_formatter(ticker.FuncFormatter(lambda y,pos: ('{{:.{:1d}f}}'.format(int(np.maximum(-np.log10(y),0)))).format(y)))
-        ax.yaxis.set_label_text("Percent Variants")
-        ax.xaxis.set_label_text("Allele Frequency")
-        """
-        plt.show()
-
-class VariantEffectCounts(object):
-
-    def __init__(self, ss, variantRDD, includeList = []):
-        """
-        Initializes a AlignmentDistribution class.
-        Computes the alignment distribution of multiple coverageRDDs.
-        :param SparkContext: the global SparkContext
-        :param alignmentRDDs: A list of bdgenomics.adam.rdd.AlignmentRDD objects
-        :param int bin_size: Division size per bin
-        """
-        if not includeList:
-            includeList = [
-                u'non_coding_transcript_exon_variant',
-                u'3_prime_UTR_variant',
-                u'5_prime_UTR_variant',
-                u'synonymous_variant',
-                u'stop_retained_variant',
-                u'conservative_inframe_insertion',
-                u'conservative_inframe_deletion',
-                u'5_prime_UTR_premature_start_codon_gain_variant',
-                u'disruptive_inframe_deletion',
-                u'disruptive_inframe_insertion',
-                u'missense_variant',
-                u'stop_lost',
-                u'start_lost',
-                u'splice_donor_variant',
-                u'splice_acceptor_variant',
-                u'splice_region_variant',
-                u'frameshift_variant',
-                u'stop_gained'
-            ]
-
-        effectCounts = variantRDD.toDF().select("annotation").rdd.map(lambda w: w[0].asDict()['transcriptEffects']) \
-            .flatMap(lambda m: m).map(lambda g: g.asDict()['effects']).flatMap(lambda a: a).filter(lambda i: i in includeList).countByValue()
-
-
-        self.sc = ss.sparkContext
-        self.effectCounts = effectCounts
-
-
-
-    def plot(self, xScaleLog = False, yScaleLog = False, testMode = False, plotType="I"):
-        """
-        Plots final distribution values and returns the plotted distribution as a counter object.
-        :param xScaleLog: rescales xaxis to log
-        :param yScaleLog: rescales yaxis to log
-        :param testMode: if true, does not generate plot. Used for testing.
-        :param plotType: Cigar type to plot, from ['I', 'H', 'D', 'M', 'S']
-        """
-        #print self.effectCounts.keys()
-
-
-        plt.rcdefaults()
-        fig, ax = plt.subplots()
-        effects = self.effectCounts.keys()
-        y_pos = np.arange(len(effects))
-        counts = [ self.effectCounts[item] for item in effects]
-        ax.barh(y_pos, counts, align='center', color='green', ecolor='black')
-        ax.set_yticks(y_pos)
-        ax.set_yticklabels(effects)
-        ax.xaxis.set_label_text("Variant Count")
-        plt.title("Variant Count By Functional Consequence")
-
-        plt.show()
-
-
-        return 1
-
